The last commit thought it was fixing a sysstat output change. However, it was incorrect and this commit reverts the fix.
The problem is the sysstat time output depends on the LANG environment
variable. for the US, it formats time as HH:MM:SS AM (or PM) while the
script does not expect the AM or PM. The correct fix is to override the
LANG variable to en_GB.UTF-8 when sar is started.