With the generation of the new dataset, I came across a very serious performance bug. What RUBBoS does is that for a story, it will fetch all root comments (level 0), and then for EVERY root comment that might have a child, it will send another query to fetch its children. As you can see, this is not scalable as for a single story, 50+ queries are sent out when only two are needed. As a result, the database will become CPU saturated very quickly.
So, I re-architected RUBBoS to perform only N queries where N is the
depth of the comment tree. The servlets then perform some post
processing to put the comments together.

A fallout of this change was that the _identical_ comment display code
found in ViewStory and ViewComment was put into one place. Negatives
are that the SQL code is slightly more hairy now.

Also look at the commented out SQL code (search for 'CASE' in the
patch). These are optimizations that fetch less data from the database
whenever the client does not need it. I did not put it in for two
reasons

(a) It changes the semantics of the benchmark. After look at the
slashcode (Slashdot's code) that RUBBoS was modeled after, I see that
they do what RUBBoS now does - fetch all then do some post
processing before displaying them
(b) Because less data is fetched, it makes it hard (if not impossible)
to compare the performance of different versions of RUBBoS